add words information print feature
For some reason, user may want to show the word information status. For
example, to know why the annoying word is showing frequently in test. It
can be done by seeing the word information file in .ago directory.
However, the file is in json format and hard to see. This commit let the
program shows the information in relatively easy-to-read style.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>