implement dictionary search feature
There are lots of ways to find out meaning of a word using computer. May
use a proprietary dictionary, dictionary on web, wikipedia, and etc.
This program, for now, will use Daum dictionary[1] search result. It is
not ideal, nor sufficient. However, reasonable and can be implemented
fast. This commit add a function to use the Daum dictionary[1] search in
this program. You can use `dic` subcommand with querying word as
argument. For example, `go run ago.go dic hello` command will shows the
meaning of the word and some example, though it is ugly and hard to read
for now.

[1] http://dic.daum.net/

Signed-off-by: SeongJae Park <sj38.park@gmail.com>