dic: add fail history of the word as it searched
Currently, as a word has been searched using dic subcommand, its
reference count becomes increased to give the word higher priority.
However, searching a word with dictionary means that user didn't
understood the word well yet. So, the searching would be similar to
failed to test the word rather than reference count increasing.
Moreover, it is more natural to give higher priority.  For the reason,
this commit change the action to add one more test fail history of the
word rather than increasing the reference count of the word.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>